Added new APIS for all objects and covered by tests. And corrected code.

+++ src/main/java/com/device/shop/service/PaymentDetailsService.java
@@ -1,6 +1,8 @@
 package com.device.shop.service;
 
+
 import com.device.shop.entity.PaymentDetails;
+import com.device.shop.exception.BadRequestException;
 import com.device.shop.repository.PaymentDetailsRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,7 @@ import javax.persistence.EntityNotFoundException;
 public class PaymentDetailsService implements PaymentDetailsServiceInterface {
 
     PaymentDetailsRepository paymentDetailsRepository;
+
     @Transactional
     public PaymentDetails createPaymentById(Long id) {
         PaymentDetails payment = new PaymentDetails(id);
@@ -22,18 +25,28 @@ public class PaymentDetailsService implements PaymentDetailsServiceInterface {
 
     @Transactional
     public PaymentDetails getPaymentById(Long paymentId) {
-        return paymentDetailsRepository.findById(paymentId)
-                .orElseThrow(() -> new EntityNotFoundException("Payment not found"));
+        return paymentDetailsRepository.findById(paymentId).orElseThrow(() -> new EntityNotFoundException("Payment not found"));
 
     }
 
     @Transactional
-    public PaymentDetails updatePaymentDetails (PaymentDetails paymentDetails) {
-        return paymentDetailsRepository.save(paymentDetails);
+    public PaymentDetails updatePaymentDetails(PaymentDetails paymentDetails, Long id) throws BadRequestException, EntityNotFoundException {
+        if (id == null || !paymentDetailsRepository.existsById(id)) {
+            throw new EntityNotFoundException("Payment Details with id " + id + " not found");
+        } else if (!id.equals(paymentDetails.getId())) {
+            throw new BadRequestException("Cannot change the id to " + paymentDetails.getId());
+        } else {
+            return paymentDetailsRepository.save(paymentDetails);
+        }
     }
+
     @Transactional
     public void deletePaymentDetailsById(Long paymentId) {
-        paymentDetailsRepository.deleteById(paymentId);
+        if (paymentDetailsRepository.existsById(paymentId)) {
+            paymentDetailsRepository.deleteById(paymentId);
+        } else {
+            throw new EntityNotFoundException("Payment with id " + paymentId + " not found");
+        }
     }
 
 }
